requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
azure-core==1.26.4
azure-storage-blob==12.16.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
fastapi==0.95.2
flower==1.2.0
h11==0.14.0
humanize==4.6.0
idna==3.4
isodate==0.6.1
kombu==5.2.4
msal==1.22.0
msal-extensions==1.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==1.10.7
pyhumps==3.8.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
